Boss and Item Updates
Boss now has art and has a separate health bar that shows up at the top of the screen. Also added some item art along with more items finished in the game.

+++ SFRL/Assets/Scripts/Items/PlayerPassive.cs
@@ -48,7 +48,7 @@ public class PlayerPassive : BasicItem
 	[Space]
 	[Header("Weapon Stats Percent")]
 	[Space]
-	public int CritChancePercentBonus;
+	public float CritChancePercentBonus;
 	public float DamagePercentBonus;
 	public float AttackSpeedPercentBonus;
 	public float RangePercentBonus;
